DF-15748 minor fix to compile latest 15.5.2 msvc and revert timeout
Former-commit-id: 6106f61a989ffc44a09f586c7548d7539966cde7

+++ Sources/Internal/UI/UIFileSystemDialog.cpp
@@ -346,7 +346,7 @@ void UIFileSystemDialog::RefreshList()
     while (true)
     {
         p = static_cast<int32>(curDirPath.rfind("/", p));
-        if (p != static_cast<int32>(curDirPath.npos))
+        if (static_cast<size_t>(p) != curDirPath.npos)
         {
             p--;
             outCnt++;
